requirements.txt
stable-baselines3>=2.3.0
gymnasium>=0.29.0
websocket-client>=1.7.0
numpy>=1.24.0
tensorboard>=2.15.0
